加入增益读取，空间不足，砍代码前备份一下

+++ Code/main.h
@@ -8,7 +8,10 @@ typedef unsigned char uint8_t;
 typedef signed char int8_t;
 typedef unsigned short uint16_t;
 typedef signed short int16_t;
-typedef unsigned int uint32_t;
-typedef signed int int32_t;
+typedef unsigned long uint32_t;
+typedef signed long int32_t;
+
+//51中int和short一样都是16位，long为32位
+//用printf输出时1byte、2byte、4byte有符号整型数据分别用bd、%hd、%ld表示
 
 #endif
